CORE-5386 Updated facepalm to allow conversions to accept a dir param.
Updated the conversion-map function to return 2 maps: the conversion functions map as well as a map of symbols of these functions.
This allows the do-conversion function to examine metadata on the functions, and to pass the unpacked conversion dir to  a conversion function if it requires the sql files.
Updated the facepalm.c200-2014042401/convert function with the required metadata so that facepalm will pass the unpacked conversion dir to it.